pyproject.toml
[project]
name = "labanquinha-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[all]>=0.115.14",
    "httpx>=0.28.1",
    "motor[snappy]>=3.7.1",
    "passlib>=1.7.4",
    "pydantic-core>=2.33.2",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pymongo>=4.13.2",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "uvicorn[standard]>=0.35.0",
]
